+++ app/src/main/java/woowacourse/app/model/MainData.kt
@@ -1,8 +1,8 @@
-package woowacourse.app.model.main
+package woowacourse.app.model
 
 import woowacourse.app.ui.main.adapter.MainViewType
 
-sealed class MainData {
+abstract class MainData {
     abstract val mainViewType: MainViewType
     abstract val id: Long
 }
