+++ app/src/main/java/woowacourse/app/model/SelectedSeatUiModel.kt
@@ -2,6 +2,7 @@ package woowacourse.app.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import woowacourse.app.model.ticket.SeatUiModel
 
 @Parcelize
 data class SelectedSeatUiModel(val selectedSeat: Set<SeatUiModel>) : Parcelable
