+++ app/src/main/java/woowacourse/app/model/advertisement/AdvertisementMapper.kt
@@ -1,4 +1,4 @@
-package woowacourse.app.model.main
+package woowacourse.app.model.advertisement
 
 import woowacourse.domain.advertisement.Advertisement
 import woowacourse.movie.R
