+++ app/src/main/java/woowacourse/app/ui/main/adapter/listview/MainViewHolder.kt
@@ -1,7 +1,7 @@
 package woowacourse.app.ui.main.adapter.listview
 
 import android.view.View
-import woowacourse.app.model.main.MainData
+import woowacourse.app.model.MainData
 import woowacourse.app.ui.main.adapter.MainViewType
 
 sealed class MainViewHolder(val view: View) {
