+++ app/src/main/java/woowacourse/app/ui/main/adapter/recyclerview/MainViewHolder.kt
@@ -2,7 +2,7 @@ package woowacourse.app.ui.main.adapter.recyclerview
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.app.model.main.MainData
+import woowacourse.app.model.MainData
 import woowacourse.app.ui.main.adapter.MainViewType
 
 sealed class MainViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
