+++ app/src/main/java/woowacourse/app/ui/main/adapter/recyclerview/MovieViewHolder.kt
@@ -4,8 +4,8 @@ import android.view.View
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
-import woowacourse.app.model.main.MainData
-import woowacourse.app.model.main.MovieUiModel
+import woowacourse.app.model.MainData
+import woowacourse.app.model.movie.MovieUiModel
 import woowacourse.app.ui.main.adapter.MainViewType
 import woowacourse.app.util.formatScreenDate
 import woowacourse.movie.R
